third_party/WebKit/LayoutTests/fast/css/parsing-not-after-supports.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<style>
span:not(.whatever) {
    background-color: red;
}
@supports (color: red) {
    span {
        color: red;
    }
}
span:not(.whatever) {
    background-color: green;
}
</style>
</head>
<body>
<span id='target'></span>
<pre id='console'></pre>
<script>
description('Test for issue 257695: CSS @supports bug: breaks following :not()');
var target = document.getElementById('target');
shouldBe('getComputedStyle(target).backgroundColor', '"rgb(0, 128, 0)"');
</script>
</body>
</html>
