third_party/blink/web_tests/platform/generic/fast/css/parsing-not-after-supports-expected.txt

Test for issue 257695: CSS @supports bug: breaks following :not()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
